+++ packages/core/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @daltonr/pathwrite-core
 
+## 0.6.2
+
 ## 0.6.1
 
 ## 0.5.0